Integration FortranMPC : Fix MPI_Attr_get.
The internal MPI_Attr_get function expects a void** pointer for the
attr_value argument.
With Fortran, we have a int* pointer.

Also make sure we always handle attr as a void**. In the MPI_Attr_put
function, a special path stores attributes coming from Fortran as
integers, not pointers to Integers. This is in conflict with default
attributes stored in defines_attr_tab.