AIO : Fix problem of POSIX compliance on old libc.
When MPC is compiled on a bugged libc relatively to the
spawing of AIO threads with large TLS, it builds its
own implementation of the AIO.

However, the aio.h file is dependent from the sigevent
struct which is POSIX.1-2001. We make sure we support it
(corret POSIX level) before including these functions.

We declare the sctk_aio functions in the mpcthread.h header (for use in ROMIO)
as they are compliant with the standard apart for signal safety.
This code section is only used on old GLIBC.