MPC_BUILD: Do not forward -Wl,-rpath through LDFLAGS but use LD_LIBRARY_PATH for dep embedding.
The issue occurs when MPC tries to load an components like SLURM or IB,
located in an external location. But, if this location points to a
default installation path (like /usr/ for example), all shared objects
in it will be included w/ RPATH. The problem occurred for system
compiler component: libgcc_s, by default installed in /usr/lib64. This
shared library sweeps our GCC installation.

It would be safe to check exhaustively that this commit does not affect third-party
components