PRIV-PLUGIN: Add multi-model guard work done (GCC 6 & 7)
This work essentially upgrades ..._per_tu_init() defs to handle
multi-model TLS (one slight change is done into each variable wrapper to
set the guard accordingly to the guard TLS model). The main function is
built following the approach below:

  1. Build N 'if' statement sequence (gimple_seq), one for each
     supported TLS level. These blocks are not directly added to the
     function body. Each if is built like this:
         * GIMPLE_ASSIGN(GUARD --> TMP_VAR)
	 * GIMPLE_COND(TMP_VAR == 0)
	 * IF_TRUE: GUARD = 1
  2. Browse variables to find which one will need a dynamic
     initialization here. Instead of directly putting the assign
     into the function body, we put this assign in the corresponding
     'if' statement, where the guard is at the same TLS model than the
     current variable. We also remember that this 'if' block contains at
     least one variable assign.
  3. Sequences with at least one assign are added to the function body,
     others are discarded. the IF_FALSE label is inserted as the last
     statement of the current block. The issue is to retrieve the
     false_label, stored into the GIMPLE_COND statement then this one
     was added at step 1 (we have to iterate over statements first).

Note: Now, if there is no need to dynamically initialize a TLS var, the
function is not inserted. This will not change the performance (function
was not called anyway) but will slightly decrease the binary size.