MPC_Config: Reset enums & structs in any case
This causes `mpc_print_config` to display everything available in the
configuration, even parameters from disabled modules. This had to be
done to avoid compiling MPC_Active_Message to make enum values available
for mpc_print_config. For sure, the best fix is to set only objects
related to enabled modules. It seems that C-structs are forced to be set
to 0, even when not set because the module is not activated while
C-enums are registered into an hash table, and failed if a lookup fails.

Files to look at for a fix :
	* sctk_runtime_config_mapper.c
	* sctk_runtime_config_defaults.c.xsl (generator)
	* sctk_runtime_config_walk.c (where the fatal() is raised)