PORTALS: Replace with PtlMESearch
Now, the sctk_thread_message_header_t is forwarded to the driver as
multiple fields have to be set when a match is found :
	- the tag, when ANY_TAG is given
	- the source rank, when ANY_SOURCE is given
	- in either case, the message size
It is clearly easier to forward the struct instead of creating a
temporary, network-related only typedef to fill data in and then copy
them back into the message header.

Please note that previously (regular probing system) the header is
memcpy'd from the message received from network and reconstructed from
the card. In that sense, the driver should already be aware of how an
MPC message is built.

This patch only support the probing for Portals eager protocol