KTHREAD: direct call to pthread_create() instead of dlsym'ing
dlsym'ing the pthread_create() symbol is a bit tricky as it may be hard for the
runtime to determine what is the link order. If the user compiles its code with
-lpthread, the library will loaded among the first and a 'RTLD_NEXT' from within
the MPC framework won't work. Also, using one of these solutions seem incomplete:
1. Using RTLD_DEFAULT, as potentially creating a loop if users are using third-party
   tools wrapping the pthreads interface.
2. dlopen'ing the 'libpthreads.so' DSO object, but I object the filename for Pthreads
   is not standard and some cases could be missed

Note this 'bug' may also arise for any library MPC is dlopen'ing (as far as now,
TBB & __libc_start_main(), as shown by git grep 'RTLD_NEXT'). This bug was not present
in earlier version (before refactoring) because the link order changed due to modularity.