MPI ABI: Modify datatypes type and constants
To comply with the ABI, the type of the datatypes has been change to
become a pointer. As a consequence, the choice has been made to adapt
the all datatype management accordingly. This commit introduces the
first changes mainly in the lowcomm.

The biggest change is the unification of datatypes. Previously, common,
contiguous and derived datatypes were differenciated. The goal is to
only have one datatype structure that is common to all datatypes. To do
so, we take the derived structure to the lowcomm.

mpc_lowcomm_datatype_t is now a pointer on this previous structure that
was renamed MPI_ABI_Datatype to match the ABI requirements. All the
constants associed have new values to match the ABI defined ones and are
typed to be mpc_lowcomm_datatype_t.

Some new datatypes has been added (REAL2 and COMPLEX4) and derived
common datatypes are stored in the same array as regular common ones.

This commit will be followed by others in which the changes will be
propagated to the upper layers of MPC mainly MPC_MPI module.