MPI ABI: Change Datatypes handle
The datatype handle is now a pointer on a generic datatype structure.
All kinds of datatypes as been merged into one generic representation.
Most of the datatype management functions have been modified to reflect
this changes.

Before, it was possible to use an uncommited type in MPI, this is not
the case anymore.
The maximum number of datatypes a user can define has been divided by
two as the contiguous datatype array has been removed (now this max is
4000).
The thread safety on accessing the datatype array has been re-thought,
the lock is now used as intended.