MPI ABI: Propagate changes in Fortran and MPI-IO
The change of datatypes handle and structure have been propagate to the
fortran and io parts of MPI.
It allows a new interface with OMP TRACE via a fortran conversion to get
an integer.
Some minor changes have been operated on the previously modified
functions and the comments have been enhanced.