MPI ABI: Fix Fortran type commit bug
The conversion from a C pointer to an Fortran int is tricky.
The integer cannot store an address surely. So, we use the id of a
datatype to perform the conversion.

The problem was uncommiteddatatypes. The conversion assumes that the datatype is stored in the
datatype array to retrieve the C structure from the Fortran integer.
An option is to store fortran created datatype in another array to
difference it from commited once. It means create another global array
just to store datatypes between creation and commit.

Another option (not fully API compliant but simpler and close to old MPC
behaviour) is to commit datatype before convert it back to fortran.
This means that uncommited datatypes can be used but it simplify greatly
the management of datatypes in Fortran.