MPI ABI: Fix the commited datatypes checks
According to the MPI Standard, commit a datatype is mandatory only for
communication. Before this commit, all use of a datatype expected it to
be commited. It resulted in broken user code as you may wish to call
infomative (MPI_Get_size...) or creation functions on uncommited
datatypes.

Now the datatype are stocked in the datatype array at creation and have
a `is_commited` field to identify valid datatypes for communication from
the others.

This change allows as well a simplier datatype management for Fortran. The
last commit that introduced implicite commit at creation has been
reverted.