MPI ABI: Fix Comm and Group in Fortran
The global value to know if we manipulate a fortran object or a C one
wasn't update at all. This leads to call the fortran inner function
(dup, delete...) of keyval with the C function and a segfault.

To fix this issue, a special value is passed to MPI_Keyval_create to set
the fortran flag to true. The execution flow is closer than it was
intended.

Other issues were due to not verifying the error code of the C calls.
It results in dereferencing pointers on errorneous structure and
possible segfault.

To fix it, we test the error code before accessing the object returned
by the C functions.