MPI ABI: Padding MPI_Status to comply with the ABI
The MPI_Status structure is likely to be defined in the standard as a
eight int (32 bytes) structure.
To correspond with this ABI, the MPC status structure has been padded
with three int that can be used for implementation specific purpose
latter on.

The position of the three standardized fields (MPI_SOURCE, MPI_TAG and
MPI_ERROR) will be specified. We think it will be as it was in MPC, the
first three int of the structure in this specific order.