MPI ABI: Change Comm handle and constants
The structure beyond the handle is now a struct MPI_ABI_Comm to comply
with the standard ABI.
The MPI_Comm constants (WORLD and SELF) are now truely constants as they
are numbers. Translation functions have been added to retrieve pointers
on those constants structure to use as any other communicator.