MPI ABI: Change Errhandler handle and constants
The handles are now adresses on MPI_ABI_Errhandler structures which are
stored in a global array.
Functions to manipulate those handles have been modified to work with
the new type.

The global array is locked for thread safety and the structure contains
a reference counter to avoid freeing in use errhandler.