OFI: support endpoint type from config
This adds a new configuration variable in the OFI driver config name eptype.
This allows to interact with the OFI provider logic to force one or another type
of Endpoint (moving from RDM to MSG for example). The type of endpoint is also
now displayed in the -v output of the mpcrun command for a better understanding
of the underlying OFI behavior.