TBSM: implementation of driver oriented (tbsm) thread communication.
First solution, which actually did not required as much dev as planned,
is to have a dedicated lcp endpoint for local task so that we can attach
a queue on each endpoint. Then, we pushing a package in the queue, only
the queue of this particular endpoint will be locked.