chore(doc): merge Doxygen code documentation in Sphinx one
The idea is to gather all the documentations (user, contributor and
developer ones) in a single place and output. With these changes, the
Doxygen only outputs xml files read by the Sphinx to incorporated the
code documentation in the output html.

This incorporation is meant to be reworked to add more levels in the
documentation when the various modules have properly formatted
documentation using grouping.

See #229