fix(MPI): remove hardcoded values and add error codes (#144)
The purpose of this commit is to remove the hardcoded errhandler values
in errh.c. The MPC_ERR_ERRHANDLER error code has been added to replace
the 17 (and other hardcoded error codes). A new macro
MPC_ERRHANDLER_NB_PREDEF is introduced to have a global view of the
number of predefined errhandler.

Closes #144