fix(lowcomm): propagate the unpacking fix to the collectives (#235)
The fix introduced to solve the unpacking issue was only applied to the
pt2pt communications. This commmit propagate the fix to the collectives
communications located within lowcomm.

This propagation implied a change of the `mpc_lowcomm_waitall` API. It
is now more alligned with the `mpc_lowcomm_wait` one.

An issue was discovered with the way request are stored and retrieved.
In this commit, a workaround is used to access directly the correct
layer of the request when located in lowcomm. The refactor of the
lowcomm API and internal objects is planned for later. This workaround
will be removed at that time.

Closes #235