fix(lcp): fix initialization allocation for multiple devices (#236)
The segfault observed in #236 is due to a wrong initialization in a
multi-nic setup. The number of devices was wrongly computed by being
limited by the maximum number of ifaces authorized by the configuration.
The amount of memory allocated for the resources is tied to the number
of devices to initialize. As a result, devices description were written
out of bounds in case of a component having multiple devices.

The fix is change of the condition of the filter loop on multi-devices
components to allow multiple devices to be tied to one component as
intended.

Closes #236