chore(MPI): add assertion on disguise for datatype unpacking
It should not be possible to be already disguised when arriving in
`_mpc_cl_request_recv_complete`. An assertion is added to the result of
`MPCX_Disguise` to confirm our assumptions. The failing assert should
point us to potential bugs in the execution (previous disguise has not
been taken off).