chore(changelog): add an unreleased section for incremental updates
This commit adds an unreleased section to the changelog and populates it
with the changes made since the last release (4.3.0). The main idea of
the addition is to introduce a way to incrementally update the changelog
along with the changes. Significant changes are supposed to be reported
into the change in one of the commit of the merge request.
Finally, the unreleased section follows keepachangelog guidelines and
should ease release process in the future.